Task 3 - News Aggregator/requirements.txt
bcrypt
mysql-connector-python
pillow
requests